app-api/requirements.txt
Django>=3.2.7,<3.3.0
djangorestframework>=3.12.4,<3.13.0
psycopg2>=2.9.1,<3.0.0
environs>=9.3.3,<9.4.0
coverage>=5.5.0,<5.6.0
pytop>=2.6.0,<2.7.0
